fix bug cancel order

+++ backend/src/order/order.service.ts
@@ -5,11 +5,25 @@ import { Order } from './order.schema';
 import { CreateOrderDto } from './dto/createOrder.dto';
 import { Request } from 'express';
 import { Product } from '../product/product.schema';
-import { statusDeliveryEnum, statusOrderEnum } from '../utils/variableGlobal';
+import {
+  paymentMethodEnum,
+  statusDeliveryEnum,
+  statusOrderEnum,
+} from '../utils/variableGlobal';
 import RequestWithRawBody from '../utils/stripe/requestWithRawBody.interface';
 import { Cart } from '../cart/cart.schema';
 import { CartIdDto } from '../cart/dto/cartId.dto';
 import { Coupon } from '../coupon/coupon.schema';
+
+interface AggregateStage {
+  $match?: any;
+  $lookup?: any;
+  $unwind?: any;
+  $project?: any;
+  $sort?: any;
+  $skip?: any;
+  $limit?: any;
+}
 
 @Injectable()
 export class OrderService {
@@ -267,7 +281,9 @@ export class OrderService {
           new: true,
         },
       );
-      await this.couponModel.findByIdAndDelete(order.coupon);
+      if (order?.coupon) {
+        await this.couponModel.findByIdAndDelete(order.coupon);
+      }
       return {
         url: session.url,
         session_id: session.id,
@@ -311,6 +327,7 @@ export class OrderService {
             {
               $set: {
                 status_delivery: statusDeliveryEnum.notShippedYet,
+                paymentMethod: paymentMethodEnum.online,
               },
             },
             {
@@ -380,25 +397,34 @@ export class OrderService {
         await findProduct.save();
       }
 
-      // eslint-disable-next-line @typescript-eslint/no-var-requires
-      const stripeWithSecretKey = require('stripe')(
-        process.env.STRIPE_SECRET_KEY ?? '',
-      );
+      if (cart.paymentMethod === paymentMethodEnum.online) {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
+        const stripeWithSecretKey = require('stripe')(
+          process.env.STRIPE_SECRET_KEY ?? '',
+        );
 
-      const refund = await stripeWithSecretKey.refunds.create({
-        payment_intent: order.payment_intent_id,
-      });
-      if (refund.status === 'succeeded') {
+        const refund = await stripeWithSecretKey.refunds.create({
+          payment_intent: order.payment_intent_id,
+        });
+        if (refund.status === 'succeeded') {
+          order.status = statusOrderEnum.cancel;
+          await order.save();
+          return {
+            msg: 'Hủy đơn hàng thành công.',
+            status: true,
+          };
+        } else {
+          return {
+            msg: 'Hủy đơn hàng không thành công.',
+            status: false,
+          };
+        }
+      } else {
         order.status = statusOrderEnum.cancel;
         await order.save();
         return {
           msg: 'Hủy đơn hàng thành công.',
           status: true,
-        };
-      } else {
-        return {
-          msg: 'Hủy đơn hàng không thành công.',
-          status: false,
         };
       }
     } catch (error) {
@@ -474,6 +500,85 @@ export class OrderService {
     }
   }
 
+  // async getAllOrdersBySeller(req: Request) {
+  //   const sellerId = req['user']._id;
+  //   const keySearch: string = req.query?.s?.toString() || '';
+  //   const currentPage: number = parseInt(req.query.page as any) || 1;
+  //   const itemsPerPage: number = parseInt(req.query.limit as any) || 10;
+
+  //   try {
+  //     const aggregateOptions: AggregateStage[] = [
+  //       { $match: { sellerId: sellerId } },
+  //       {
+  //         $lookup: {
+  //           from: 'Users',
+  //           localField: 'userId',
+  //           foreignField: '_id',
+  //           as: 'user',
+  //         },
+  //       },
+  //       {
+  //         $lookup: {
+  //           from: 'Products',
+  //           localField: 'productId',
+  //           foreignField: '_id',
+  //           as: 'product',
+  //         },
+  //       },
+  //       { $unwind: '$user' },
+  //       { $unwind: '$product' },
+  //       {
+  //         $match: {
+  //           'product.name': { $regex: new RegExp(keySearch, 'i') },
+  //           'user.username': { $regex: new RegExp(keySearch, 'i') },
+  //           status_delivery: {
+  //             $in: [
+  //               statusDeliveryEnum.notShippedYet,
+  //               statusDeliveryEnum.shipping,
+  //               statusDeliveryEnum.shipped,
+  //               statusDeliveryEnum.cancel,
+  //             ],
+  //           },
+  //         },
+  //       },
+  //       {
+  //         $project: {
+  //           cartId: '$_id',
+  //           image: '$product.variants.image',
+  //           sold: '$product.variants.sold',
+  //           color: '$product.variants.color',
+  //           price: '$product.variants.price',
+  //           quantity: 1,
+  //           status_delivery: 1,
+  //           buyername: '$user.username',
+  //           productName: '$product.name',
+  //           variantName: '$product.variants.color',
+  //           createdAt: 1,
+  //           updatedAt: 1,
+  //         },
+  //       },
+  //       { $sort: { createdAt: -1 } },
+  //       { $skip: (currentPage - 1) * itemsPerPage },
+  //       { $limit: itemsPerPage },
+  //     ];
+
+  //     const totalOrders = await this.cartModel
+  //       .aggregate(aggregateOptions)
+  //       .count('total');
+  //     const orders = await this.cartModel.aggregate(aggregateOptions);
+
+  //     return {
+  //       status: true,
+  //       orders,
+  //       totalOrders,
+  //       page: currentPage,
+  //       limit: itemsPerPage,
+  //     };
+  //   } catch (error) {
+  //     throw new BadRequestException(error);
+  //   }
+  // }
+
   async paymentOrderWithCOD(req: RequestWithRawBody) {
     try {
       const order = (await this.getUserOrderPending(req))?.order;
@@ -495,6 +600,7 @@ export class OrderService {
         {
           $set: {
             status_delivery: statusDeliveryEnum.notShippedYet,
+            paymentMethod: paymentMethodEnum.offline,
             createdAt: new Date(),
           },
         },
